+++ src/Video/Models/VideoProxy.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Video\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class VideoProxy extends ModelProxy
+{
+}