+++ src/Video/resources/database/migrations/2025_03_05_120417_create_model_videos_table.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up(): void
+    {
+        Schema::create('model_videos', static function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('video_id')->constrained('videos')->onDelete('cascade');
+            $table->morphs('model');
+            $table->integer('priority')->default(0);
+            $table->boolean('is_primary')->default(false);
+            $table->timestamps();
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::dropIfExists('model_videos');
+    }
+};