+++ src/Video/resources/manifest.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'name' => 'Vanilo Video Module',
+    'version' => '5.0-dev'
+];